+++ win/win32/mhtext.c
@@ -53,7 +53,7 @@ HWND mswin_init_text_window () {
 	SetWindowText(ret, "Text");
 	if( !GetNHApp()->bWindowsLocked ) {
 		DWORD style;
-		style = GetWindowLongPtr(ret, GWL_STYLE);
+		style = (DWORD)GetWindowLongPtr(ret, GWL_STYLE);
 		style |= WS_CAPTION;
 		SetWindowLongPtr(ret, GWL_STYLE, style);
 		SetWindowPos(ret, NULL, 0, 0, 0, 0, SWP_NOMOVE | SWP_NOSIZE | SWP_NOZORDER | SWP_FRAMECHANGED);
